requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Deep Learning
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter
torch-sparse

# Materials Science
ase>=3.22.0
pymatgen>=2023.5.0
matminer>=0.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Experiment Management
wandb>=0.15.0
mlflow>=2.4.0
hydra-core>=1.3.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
